+++ app/Http/Controllers/head/HeadTeacherController.php
@@ -8,6 +8,7 @@ use App\Models\Teacher;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Validator;
 
@@ -97,10 +98,42 @@ class HeadTeacherController extends Controller
         return view('head.teacher.index',compact('teachers'));
     }
 
-    public function delete(User $teacher){
+    public function delete(User $teacher,Request $req){
+        $checkTeacherOnClass = DB::table('mapping_class_teachers as mct')
+                                    ->leftJoin('class_transactions as ct','ct.id','mct.class_id')
+                                    ->where('ct.is_freeze','!=',1)
+                                    ->where('mct.user_id',$teacher->id)
+                                    ->first();
+        if(is_null($checkTeacherOnClass)){
+            $delete = User::find($teacher->id);
+            $delete->delete();
+            return redirect()->back()->with('msg','Success Delete Data Teacher');
+        } else {
+            $keyword=$req->search;
+            $teachers = User::where('role','teacher')
+                            ->where('id','!=',$teacher->id)
+                            ->where(function($q) use ($keyword){
+                                if(!is_null($keyword)){
+                                    $q->where('name','like',"%$keyword%");
+                                } 
+                            })
+                            ->orderBy('id','desc')
+                            ->paginate(5);
+            return view('head.teacher.teacherSwitch',compact('teachers','teacher'));
+        }
+    }
+
+    public function replace(User $teacher,$replaceTeacherID){
+        DB::table('mapping_class_teachers as mct')
+                                    ->leftJoin('class_transactions as ct','ct.id','mct.class_id')
+                                    ->where('ct.is_freeze','!=',1)
+                                    ->where('mct.user_id',$teacher->id)
+                                    ->update([
+                                        'mct.user_id' => $replaceTeacherID
+                                    ]);
         $delete = User::find($teacher->id);
-        $delete->delete();
-        return redirect()->back()->with('msg','Success Delete Teacher');
+        $delete->delete();                       
+        return redirect()->route("headTeacherPage")->with('msg','Success Replace & Delete Data Teacher');
     }
